peer edit meghan siritzky
Overall I think you're asking some really interesting questions, and you did a great job tidying the data to be able to more meaningfully interpret them. I was also impressed by the breadth of content you were able to capture in the plots. It's clear that to be able to tell the whole story, you need to show the relations between multiple variables at a time, and you did a great job of using graphs to do that. Moreover, I appreciated that your plots built on each other, and each added layer of complexity made sense coming from the previous plot.

Great use of captions and tags in ggplot! I didn't know you could use tags to identify the order of your plots - I'll use that in the future.

Broadly speaking, the areas of improvement I'd focus on are a) labeling your plots more clearly, and b) playing around with your geom_smooth plot. There's nothing wrong with it at all, I'm just not sure it is the clearest format for being able to look at that specific information. I put a potential alternative - a geom_area plot. I think it's just a question of what kind of graph you think will best depict the data you want to look at.